+++ core/src/main/java/com/arnyminerz/escalaralcoiaicomtat/core/ui/element/climb/SectorsExplorer.kt
@@ -1,0 +1,29 @@
+package com.arnyminerz.escalaralcoiaicomtat.core.ui.element.climb
+
+import android.app.Activity
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import coil.annotation.ExperimentalCoilApi
+import com.arnyminerz.escalaralcoiaicomtat.core.ui.viewmodel.SectorsViewModel
+
+@Composable
+@ExperimentalMaterialApi
+@ExperimentalAnimationApi
+@ExperimentalCoilApi
+@ExperimentalFoundationApi
+fun SectorsExplorer(
+    activity: Activity,
+    navController: NavController,
+    areaId: String,
+    zoneId: String
+) {
+    Explorer(
+        activity,
+        navController,
+        dataClassViewModel = SectorsViewModel::class.java,
+        viewModelArguments = listOf(areaId, zoneId)
+    )
+}
